fix: ask-handy-button
